Add public documentation
