+++ backend/documentation/public-swagger.json
@@ -1,0 +1,113 @@
+{
+  "swagger": "2.0",
+  "info": {
+    "description": "Documentation for the public endpoints supported by the Monketype API server.\n\nNote that authentication is performed with the Authorization HTTP header in the format `Authorization: ApeKey YOUR_APE_KEY` ",
+    "version": "1.0.0",
+    "title": "Monkeytype API",
+    "termsOfService": "https://monkeytype.com/terms-of-service",
+    "contact": {
+      "name": "Developer",
+      "email": "jack@monkeytype.com"
+    }
+  },
+  "basePath": "/",
+  "consumes": ["application/json"],
+  "produces": ["application/json"],
+  "tags": [
+    {
+      "name": "users",
+      "description": "User data and related operations"
+    }
+  ],
+  "paths": {
+    "/users/personalBests": {
+      "get": {
+        "tags": ["users"],
+        "summary": "Gets a user's personal best data",
+        "parameters": [
+          {
+            "name": "mode",
+            "in": "query",
+            "description": "The primary mode (i.e., time)",
+            "required": true,
+            "type": "string"
+          },
+          {
+            "name": "mode2",
+            "in": "query",
+            "description": "The secondary mode (i.e., 60)",
+            "required": false,
+            "type": "string"
+          }
+        ],
+        "responses": {
+          "200": {
+            "description": "",
+            "schema": {
+              "$ref": "#/definitions/PersonalBest"
+            }
+          }
+        }
+      }
+    }
+  },
+  "definitions": {
+    "Response": {
+      "type": "object",
+      "required": ["message", "data"],
+      "properties": {
+        "message": {
+          "type": "string"
+        },
+        "data": {
+          "type": "object"
+        }
+      }
+    },
+    "PersonalBest": {
+      "type": "object",
+      "properties": {
+        "acc": {
+          "type": "number",
+          "format": "double",
+          "example": 94.44
+        },
+        "consistency": {
+          "type": "number",
+          "format": "double",
+          "example": 75.98
+        },
+        "difficulty": {
+          "type": "string",
+          "example": "normal"
+        },
+        "lazyMode": {
+          "type": "boolean",
+          "example": false
+        },
+        "language": {
+          "type": "string",
+          "example": "english"
+        },
+        "punctuation": {
+          "type": "boolean",
+          "example": false
+        },
+        "raw": {
+          "type": "number",
+          "format": "double",
+          "example": 116.6
+        },
+        "wpm": {
+          "type": "number",
+          "format": "double",
+          "example": 107.6
+        },
+        "timestamp": {
+          "type": "integer",
+          "example": 1644438189583
+        }
+      }
+    }
+  }
+}
